ALPHA 1.5 V

+++ EmployeeProject/src/employee/version2/PieceWorkerEmployee.java
@@ -40,7 +40,7 @@ public class PieceWorkerEmployee extends Employee {
         this.totalPiecesFinished = totalPiecesFinished;
     }
     
-    public void setRatePerPiece(float ratePerPiece) {
+    public void setRatePerPiece(double ratePerPiece) {
         this.ratePerPiece = ratePerPiece;
     }
